refactor(caste-client-ng): moved user calls into users.service as oposed to user.service.ts
