+++ projects/caste-client-ng/src/lib/management/caste-user.service.ts
@@ -20,15 +20,9 @@ import { NewAccount } from '../interfaces/new_account.interface';
   providedIn: 'root',
 })
 export class CasteUserService extends BaseService {
-  public httpClient: HttpClient;
   public user: User;
   public selectedAccount: Account;
   public application: Application;
-
-  constructor(@Inject(CASTE_AUTH_CONFIG) config: CasteAuthConfig, http: HttpClient) {
-    super(http, { ...DEFAULT_CONFIG, ...config });
-    this.httpClient = http;
-  }
 
   public setUser(user: User) {
     this.user = user;
@@ -89,56 +83,5 @@ export class CasteUserService extends BaseService {
 
   public getToken(): string {
     return localStorage.getItem(this.opts.tokenStorageKey);
-  }
-
-  public verifyEmail(userId: string, token: string) {
-    return this.httpClient.put(
-      `${this.opts.url}/public/users/${userId}/verify`,
-      {
-        realm: this.opts.realm,
-        code: token,
-      },
-      {
-        headers: this.opts.baseHeaders,
-      },
-    );
-  }
-
-  public requestPasswordRecovery(username: string) {
-    return this.httpClient
-      .post(
-        `${this.opts.url}/public/users/recover`,
-        { username, realm: this.opts.realm },
-        {
-          headers: this.opts.baseHeaders,
-        },
-      )
-      .pipe(map(this.extractData), catchError(this.handleError.bind(this)));
-  }
-
-  public recoverPassword(userId: string, code: string, password: string) {
-    return this.httpClient
-      .put(
-        `${this.opts.url}/public/users/${userId}/recover`,
-        { code, password, realm: this.opts.realm },
-        {
-          headers: this.opts.baseHeaders,
-        },
-      )
-      .pipe(map(this.extractData), catchError(this.handleError.bind(this)));
-  }
-
-  public createNewAccount(data: NewAccount) {
-    return this.post<NewAccount>('/user/new_account', {
-      realm: this.opts.realm,
-      ...data,
-    });
-  }
-
-  public givePermissionsToAccount(data: GivePermissions) {
-    return this.post<GivePermissions>('/user/permissions', {
-      realm: this.opts.realm,
-      ...data,
-    });
   }
 }
